Docs: Renamed file name
